Oops, forgot to commit.
Progress so far:
- ask for permission to read external storage
- allow user to select video
- read video into a MediaMetadataRetriever
- send one frame from each second of the video to the MLKit processor

Here's where things get odd. It doesn't detect any faces.
If I load in a static image screenshotted from the first second of the video then it works just fine.

Next step is to try and address this informational message, because I think that might be something to do with it:

"ML Kit has detected that you seem to pass camera frames to the detector as a Bitmap object. This is inefficient. Please use YUV_420_888 format for camera2 API or NV21 format for (legacy) camera API and directly pass down the byte array to ML Kit."