Implement token refreshing
The youtube class has been reworked to allow the getAuthentication method to do a http request to refresh the access token.

This means that a callback is required fore the getAuthentication method, and the preExecute method new needs to be suspendable until that callback returns.

In addition, auth is collected directly before the youtube api request runs to ensure as little time passes as possible between checking the refresh time and executing the request.